backend/requirements.txt
django
djangorestframework
cors
django-cors-headers
pyjwt